+++ misc/changelog.md
@@ -12,5 +12,6 @@ Changelog:
 | 8.12   | Revert back to basic Docker Postgres for testing: compatible for regular and e2e-Playwright tests. |
 | 10.12  | Add frontend unit tests for: Login, Welcome -components. Configure Vitest, CI. Add coverage for front and backend via Codecov. |
 | 12.12  | Add file upload functionality to backend and frontend. Add logout with session deletion. Uploads are now in /uploads/user.id/file.xlsx |
-| 16.12  | Add backend tests for file upload and logout.  |
+| 16.12  | Add backend tests for file upload and logout. |
+| 17.12  | Add e2e tests for file upload, refactor e2e tests. Add test-uploads dir for test files and deletion of them afterwards. |